requirements.txt
discord.py==2.3.2
langchain==0.1.0
langchain_openai==0.0.2
openai==1.6.1
urllib3==1.26.18
